[modify] 重複コードの削除
