+++ app/assets/stylesheets/cms/_addon_file.scss
@@ -29,5 +29,6 @@
   .cms-addon-file-selected-files {
     display: flex;
     flex-wrap: wrap;
+    gap: 20px;
   }
 }
