+++ app/assets/stylesheets/ss/_pc_mb.scss
@@ -2072,7 +2072,7 @@ footer.send {
   }
   #selected-files {
     margin-top: 10px;
-    @include clear;
+    //@include clear;
     @include mb {
       margin-top: 10px;
     }
@@ -2081,14 +2081,15 @@ footer.send {
 #selected-files, .user-files {
   @include display-flex;
   @include flex-wrap(wrap);
-  clear: both;
+  gap: 20px;
+  //clear: both;
 }
 .file-view {
   position: relative;
   width: 164px;
-  margin: 0 28px 20px 0;
+  //margin: 0 28px 20px 0;
   @include mb {
-    margin: 0 10px 10px 0;
+    //margin: 0 10px 10px 0;
   }
   .sanitizer-status {
     position: absolute;
@@ -2098,7 +2099,7 @@ footer.send {
     min-height: 110px;
     overflow: hidden;
     border: 5px solid $gray4;
-    line-height: 110px;
+    //line-height: 110px;
     text-align: center;
     &:hover {
       background-color: #eee;
@@ -2119,11 +2120,18 @@ footer.send {
       position: absolute;
     }
     img {
-      max-width: 120px;
-      vertical-align: middle;
+      padding: 10px;
+      width: 140px;
+      height: 110px;
+      object-fit: scale-down;
     }
     .ext {
-      display: block;
+      display: inline-flex;
+      align-items: center;
+      justify-content: center;
+      padding: 10px;
+      width: 140px;
+      height: 110px;
       overflow: hidden;
       color: $orange;
       font-size: $px32;
@@ -2131,19 +2139,22 @@ footer.send {
     }
   }
   .name {
-    padding: 10px 0;
+    padding: 5px;
     overflow: hidden;
     font-size: 90%;
     @include ellipsis;
-    word-wrap: nowrap;
   }
   .action {
-    display: inline-block;
-    margin-bottom: 5px;
+    display: inline-flex;
+    align-items: center;
+    gap: 5px;
+    flex-wrap: wrap;
     a {
       @include btn;
-      margin-bottom: 5px;
     }
+  }
+  * ~ .action {
+    margin-top: 10px;
   }
 }
