+++ app/components/ss/file_field_component.rb
@@ -4,7 +4,7 @@ class SS::FileFieldComponent < ApplicationComponent
   include ActiveModel::Model
 
   attr_accessor :object_name, :object_method
-  attr_writer :ss_mode, :cur_site, :cur_user, :cur_node, :item, :file
+  attr_writer :ss_mode, :cur_site, :cur_user, :cur_node, :item, :file, :element_id
 
   delegate :sanitizer_status, to: :helpers
 
@@ -29,11 +29,7 @@ class SS::FileFieldComponent < ApplicationComponent
   end
 
   def file
-    @file = item.send(object_method)
-  end
-
-  def sanitize_to_id(name)
-    name.to_s.delete("]").tr("^-a-zA-Z0-9:.", "_")
+    @file ||= item.send(object_method)
   end
 
   def element_id
@@ -45,13 +41,19 @@ class SS::FileFieldComponent < ApplicationComponent
       case ss_mode
       when :cms
         if cur_node
-          view_context.cms_apis_node_temp_files_path
+          view_context.cms_apis_node_temp_files_path(site: cur_site, cid: cur_node)
         else
-          view_context.cms_apis_temp_files_path
+          view_context.cms_apis_temp_files_path(site: cur_site)
         end
       else
         view_context.sns_apis_temp_files_path(user: cur_user)
       end
     end
+  end
+
+  private
+
+  def sanitize_to_id(name)
+    name.to_s.delete("]").tr("^-a-zA-Z0-9:.", "_")
   end
 end
