--- app/views/cms/frames/temp_files/files/_form.html.erb
@@ -1,65 +1,0 @@
-<dl class="see">
-  <dt><%= @model.t :name %><%= @model.tt :name %></dt>
-  <dd><%= f.text_field :name %></dd>
-
-  <dt><%= @model.t :filename %><%= @model.tt :filename %></dt>
-  <dd><%= f.text_field :filename %></dd>
-
-  <dt><%= @model.t :geo_location %><%= @model.tt :geo_location %></dt>
-  <dd>
-    <%= number_field_tag "#{f.object_name}[geo_location][lat]", @item.geo_location.try(:lat), step: :any %>
-    <%= number_field_tag "#{f.object_name}[geo_location][lng]", @item.geo_location.try(:lng), step: :any %>
-  </dd>
-</dl>
-
-<% if @item.persisted? && @item.image? %>
-  <div class="ss-image-edit-canvas">
-    <div class="toolbar">
-      <div class="btn-group">
-        <button type="button" class="btn btn-zoom-in" data-func="zoomIn" aria-label="<%= t("ss.buttons.zoom_in") %>"><%= md_icons.filled "zoom_in", tag: :i %></button>
-        <button type="button" class="btn btn-zoom-out" data-func="zoomOut" aria-label="<%= t("ss.buttons.zoom_out") %>"><%= md_icons.filled "zoom_out", tag: :i %></button>
-      </div>
-      <div class="btn-group">
-        <button type="button" class="btn btn-rotate-left" data-func="rotateRight"><%= md_icons.filled "rotate_left", tag: :i %></button>
-        <button type="button" class="btn btn-rotate-right" data-func="rotateLeft"><%= md_icons.filled "rotate_right", tag: :i %></button>
-      </div>
-      <div class="btn-group">
-        <button type="button" class="btn btn-submit-crop" data-func="submit" aria-label="<%= t("ss.buttons.send") %>"><%= md_icons.filled "done", tag: :i %></button>
-        <button type="button" class="btn btn-cancel-crop" data-func="cancel" aria-label="<%= t("ss.buttons.cancel") %>"><%= md_icons.filled "close", tag: :i %></button>
-      </div>
-
-      <div class="btn-group">
-        <div class="input-group">
-          <%= select_tag('size', options_for_select(@model.resizing_options(user: @cur_user, node: cur_node)), include_blank: true) %>
-        </div>
-      </div>
-
-      <div class="btn-group">
-        <div class="input-group">
-          <label for="ss-image-edit-canvas-inspection-x">X</label>
-          <input name="x" id="ss-image-edit-canvas-inspection-x" type="text" readonly>
-        </div>
-        <div class="input-group">
-          <label for="ss-image-edit-canvas-inspection-y">Y</label>
-          <input name="y" id="ss-image-edit-canvas-inspection-y" type="text" readonly>
-        </div>
-        <div class="input-group">
-          <label for="ss-image-edit-canvas-inspection-width">W</label>
-          <input name="width" id="ss-image-edit-canvas-inspection-width" type="text" readonly>
-        </div>
-        <div class="input-group">
-          <label for="ss-image-edit-canvas-inspection-height">H</label>
-          <input name="height" id="ss-image-edit-canvas-inspection-height" type="text" readonly>
-        </div>
-      </div>
-    </div>
-    <img src="<%= @item.no_cache_url %>" class="target" data-format="<%= @item.content_type %>">
-    <%= f.hidden_field :in_data_url, value: nil %>
-
-    <div class="confirm" style="visibility: hidden;"><%= t("ss.confirm.submit_image_canvas") %></div>
-  </div>
-
-  <%= jquery do %>
-    var imageEditor = new SS_ImageEditor('.ss-image-edit-canvas');
-  <% end %>
-<% end %>
