+++ app/views/cms/frames/temp_files/files/edit.html.erb
@@ -1,20 +1,1 @@
-<%= form_with scope: :item, url: url_for(action: :update, file_size: params[:file_size]), id: "ajax-form", method: :put, data: { turbo: true } do |f| %>
-  <%= hidden_field_tag :single, params[:single] if params[:single].present? %>
-  <% params.dig(:s, :types).tap do |types| %>
-    <% if types %>
-      <% types.each do |type| %>
-        <%= hidden_field_tag "s[types][]", type %>
-      <% end %>
-    <% end %>
-  <% end %>
-  <%= hidden_field_tag "s[node_bound]", params.dig(:s, :node_bound) %>
-  <%= hidden_field_tag "s[keyword]", params.dig(:s, :keyword) %>
-
-  <%= error_messages_for :item %>
-  <%= render 'form', f: f %>
-
-  <footer class="send">
-    <%= f.submit t("ss.buttons.save"), class: "btn-primary save", data: { disable_with: t("ss.buttons.saving") } %>
-    <%= link_to(t("ss.buttons.cancel"), url_for(action: :index, s: params[:s].try(:to_unsafe_h)), class: %w(btn-default), data: { turbo: true }) %>
-  </footer>
-<% end %>
+<%= render template: "sns/frames/temp_files/files/edit" %>
