+++ app/views/sns/frames/temp_files/files/index.html.erb
@@ -3,88 +3,4 @@
   <%= link_to tag.span(t("ss.buttons.select_from_list"), class: "tab-name"), sns_frames_temp_files_files_path(user: @cur_user, s: params[:s].try(:to_unsafe_h)), class: "current", data: { turbo: true } %>
 </div>
 
-<div class="cms-temp-file" data-controller="ss--temp-file">
-  <div class="search-ui-form">
-    <%= form_with scope: :s, url: url_for(action: :index), method: :get, class: "search", data: { turbo: true } do |f| %>
-      <div>
-        <label>
-          <%= check_box_tag "#{f.object_name}[types][]", "temp_file", @s.types && @s.types.include?("temp_file"), id: nil, data: { controller: "ss--submit-on-change" } %>
-          <%= t("mongoid.models.ss/temp_file") %>
-        </label>
-        <label>
-          <%= check_box_tag "#{f.object_name}[types][]", "user_file", @s.types && @s.types.include?("user_file"), id: nil, data: { controller: "ss--submit-on-change" } %>
-          <%= t("mongoid.models.ss/user_file") %>
-        </label>
-      </div>
-      <%= hidden_field_tag :single, params[:single] if params[:single].present? %>
-      <%= hidden_field_tag :cid, params[:cid] if params[:cid].present? %>
-      <%= f.text_field :keyword %>
-      <%= f.submit  t("ss.buttons.search"), class: :btn %>
-    <% end %>
-  </div>
-
-  <div class="user-files">
-    <% items.each do |item| %>
-      <%
-        item = item.becomes_with_model
-        if item.is_a?(SS::TempFile)
-          item = item.becomes_with(Cms::TempFile)
-        end
-
-        data_hash = Hash[%i[id name humanized_name extname url image? thumb_url image_dimension].map { |m| [m, (item.send(m) rescue nil)] }]
-        data_hash[:file_id] = item.id
-        data_hash[:size] = number_to_human_size(item.size) rescue nil
-        data_hash[:updated] = item.updated.to_i rescue nil
-        data_hash[:user_name] = item.user.name if item.user.present?
-        if item.sanitizer_state
-          data_hash[:sanitizer_state] = item.sanitizer_state
-          data_hash[:sanitizer_state_label] = item.label(:sanitizer_state)
-        end
-
-        case item
-        when Cms::File
-          data_hash[:type] = 'cms_file'
-        when SS::UserFile
-          data_hash[:type] = 'user_file'
-        else # SS::TempFile, Cms::TempFile
-          data_hash[:type] = 'temp_file'
-        end
-      %>
-      <%= content_tag(:article, class: "file-view", id: "file-#{item.id}", data: data_hash) do %>
-        <%= sanitizer_status(item) %>
-        <%
-          opts = { action: :select, id: item }
-          opts[:file_size] = true if params[:file_size].present?
-          opts[:owner_item_id] = params[:owner_item_id] if params[:owner_item_id].present?
-          opts[:owner_item_type] = params[:owner_item_type] if params[:owner_item_type].present?
-        %>
-        <%= link_to "#", class: "select", title: item.humanized_name, data: { humanized_name: item.humanized_name, action: "ss--temp-file#selectFile" } do %>
-          <div class="thumb">
-            <% if item.image? %>
-              <img src="<%= "#{item.thumb_url}?_=#{item.updated.to_i}" %>" alt="<%= item.humanized_name %>"/>
-            <% else %>
-              <span class="ext icon-<%= item.extname %>"><%= item.extname %></span>
-            <% end %>
-          </div>
-          <div class="name">
-            <%= item.name %>
-            <% if params[:file_size] %>
-              <br/>
-              <span class="file-size"><%= "( #{number_to_human_size(item.size)} )" %></span>
-            <% end %>
-          </div>
-        <% end %>
-        <nav class="menu">
-          <% if item.allowed?(:edit, @cur_user) %>
-            <%= link_to t("ss.buttons.edit"), url_for(action: :edit, id: item, s: params[:s].try(:to_unsafe_h)), class: "btn", data: { turbo: true } %>
-          <% end %>
-          <% if item.allowed?(:delete, @cur_user) %>
-            <%= button_tag t("ss.buttons.delete"), type: :button, name: :delete, class: "btn", data: { href: url_for(action: :destroy, id: item, s: params[:s].try(:to_unsafe_h), format: :json), action: "ss--temp-file#deleteFile" } %>
-          <% end %>
-        </nav>
-      <% end %>
-    <% end %>
-  </div>
-
-  <%= paginate items if items.try(:current_page) %>
-</div>
+<%= render "sns/frames/temp_files/files/index" %>
